Upgrade YamlDotNet to 16.1.3 and remove workarounds no longer needed - Remove workarounds related to being able to deserialize nested objects, as it's now passed into converter APIs - Created `NamedObjectYamlConverter` to simplify generic type conversions - Created `NamedObjectMappingYamlConverter` to handle (de)serialization of `NamedObjectMapping{TValue}`, replacing the previous implementation of `IYamlConvertible`

+++ src/Persistence/PaYaml/Models/PaYamlLocation.cs
@@ -5,7 +5,7 @@ using YamlDotNet.Core;
 
 namespace Microsoft.PowerPlatform.PowerApps.Persistence.PaYaml.Models;
 
-public record PaYamlLocation(int Line, int Column)
+public record PaYamlLocation(long Line, long Column)
 {
     internal static PaYamlLocation? FromMark(Mark mark)
     {
